CSAT-36 Refactor: Add Message Builder class to OpenAI model service.
This commit adds a Message Builder class to the OpenAI model service. This class is responsible for building the message that will be sent to the OpenAI API. The logic of building the message is now encapsulated in this class.

Furthermore, the logic of the `/message/generate` endpoint has been moved completely to the Model Service client class. This makes the endpoint code much cleaner and easier to read.